+++ crates/gqlitedb/src/tests/evaluators.rs
@@ -9,7 +9,7 @@ use crate::{
 fn test_evaluate_simple_create_node()
 {
   let temp_file = crate::tests::create_tmp_file();
-  let store = crate::store::redb::Store::new(temp_file.path()).unwrap();
+  let store = crate::store::redb::Store::open(temp_file.path()).unwrap();
 
   eval_program(&store, &programs::simple_create(), &Default::default()).unwrap();
   check_stats(&store, None, 1, 0, 0, 0);
@@ -19,7 +19,7 @@ fn test_evaluate_simple_create_node()
 fn test_evaluate_create_named_node()
 {
   let temp_file = crate::tests::create_tmp_file();
-  let store = crate::store::redb::Store::new(temp_file.path()).unwrap();
+  let store = crate::store::redb::Store::open(temp_file.path()).unwrap();
 
   let value = eval_program(&store, &programs::create_named_node(), &Default::default()).unwrap();
   check_stats(&store, None, 1, 0, 0, 1);
@@ -34,7 +34,7 @@ fn test_evaluate_create_named_node()
 fn test_evaluate_create_named_node_double_return()
 {
   let temp_file = crate::tests::create_tmp_file();
-  let store = crate::store::redb::Store::new(temp_file.path()).unwrap();
+  let store = crate::store::redb::Store::open(temp_file.path()).unwrap();
 
   let value = eval_program(
     &store,
@@ -54,7 +54,7 @@ fn test_evaluate_create_named_node_double_return()
 fn test_evaluate_double_with_return()
 {
   let temp_file = crate::tests::create_tmp_file();
-  let store = crate::store::redb::Store::new(temp_file.path()).unwrap();
+  let store = crate::store::redb::Store::open(temp_file.path()).unwrap();
 
   let value = eval_program(&store, &programs::double_with_return(), &Default::default()).unwrap();
   check_stats(&store, None, 0, 0, 0, 0);
@@ -66,7 +66,7 @@ fn test_evaluate_double_with_return()
 fn test_evaluate_unwind()
 {
   let temp_file = crate::tests::create_tmp_file();
-  let store = crate::store::redb::Store::new(temp_file.path()).unwrap();
+  let store = crate::store::redb::Store::open(temp_file.path()).unwrap();
 
   let value = eval_program(&store, &programs::unwind(), &Default::default()).unwrap();
   check_stats(&store, None, 0, 0, 0, 0);
@@ -78,7 +78,7 @@ fn test_evaluate_unwind()
 fn test_evaluate_match_loop()
 {
   let temp_file = crate::tests::create_tmp_file();
-  let store = crate::store::redb::Store::new(temp_file.path()).unwrap();
+  let store = crate::store::redb::Store::open(temp_file.path()).unwrap();
 
   let node = graph::Node {
     key: graph::Key { uuid: 1 },
@@ -118,7 +118,7 @@ fn test_evaluate_match_loop()
 fn test_evaluate_optional_match()
 {
   let temp_file = crate::tests::create_tmp_file();
-  let store = crate::store::redb::Store::new(temp_file.path()).unwrap();
+  let store = crate::store::redb::Store::open(temp_file.path()).unwrap();
 
   let value = eval_program(&store, &programs::optional_match(), &Default::default()).unwrap();
   check_stats(&store, None, 0, 0, 0, 0);
@@ -133,7 +133,7 @@ fn test_evaluate_optional_match()
 fn test_evaluate_match_count()
 {
   let temp_file = crate::tests::create_tmp_file();
-  let store = crate::store::redb::Store::new(temp_file.path()).unwrap();
+  let store = crate::store::redb::Store::open(temp_file.path()).unwrap();
   let function_manager = functions::Manager::new();
   let program = programs::match_count(&function_manager);
 
@@ -172,7 +172,7 @@ fn test_evaluate_match_count()
 fn test_evaluate_aggregation()
 {
   let temp_file = crate::tests::create_tmp_file();
-  let store = crate::store::redb::Store::new(temp_file.path()).unwrap();
+  let store = crate::store::redb::Store::open(temp_file.path()).unwrap();
   let function_manager = functions::Manager::new();
   let program = programs::aggregation(&function_manager);